+++ source/Internal/Framework/Module/Install/Service/ModuleConfigurationInstallerInterface.php
@@ -14,9 +14,13 @@ interface ModuleConfigurationInstallerInterface
     /**
      * @param string $moduleSourcePath
      * @param string $moduleTargetPath
-     * @return
      */
-    public function install(string $moduleSourcePath, string $moduleTargetPath);
+    public function install(string $moduleSourcePath, string $moduleTargetPath): void;
+
+    /**
+     * @param string $modulePath
+     */
+    public function uninstall(string $modulePath): void;
 
     /**
      * @param string $packagePath